Move FtpData classes into separate files
