+++ Sming/SmingCore/Network/Ftp/FtpServerConnection.cpp
@@ -9,166 +9,12 @@
  ****/
 
 #include "FtpServerConnection.h"
-#include "../FtpServer.h"
-#include "NetUtils.h"
-#include "TcpConnection.h"
-#include "FileSystem.h"
-
-class FtpDataStream : public TcpConnection
-{
-public:
-	explicit FtpDataStream(FtpServerConnection* connection) : TcpConnection(true), parent(connection)
-	{
-	}
-
-	err_t onConnected(err_t err) override
-	{
-		//response(125, "Connected");
-		setTimeOut(300); // Update timeout
-		return TcpConnection::onConnected(err);
-	}
-
-	err_t onSent(uint16_t len) override
-	{
-		sent += len;
-		if(written < sent || !completed) {
-			return TcpConnection::onSent(len);
-		}
-		finishTransfer();
-		return TcpConnection::onSent(len);
-	}
-
-	void finishTransfer()
-	{
-		close();
-		parent->dataTransferFinished(this);
-	}
-
-	void response(int code, String text = nullptr)
-	{
-		parent->response(code, text);
-	}
-
-	int write(const char* data, int len, uint8_t apiflags = 0)
-	{
-		written += len;
-		return TcpConnection::write(data, len, apiflags);
-	}
-
-	void onReadyToSendData(TcpConnectionEvent sourceEvent) override
-	{
-		if(!parent->isCanTransfer()) {
-			return;
-		}
-		if(completed && written == 0) {
-			finishTransfer();
-		}
-		transferData(sourceEvent);
-	}
-
-	virtual void transferData(TcpConnectionEvent sourceEvent)
-	{
-	}
-
-protected:
-	FtpServerConnection* parent = nullptr;
-	bool completed = false;
-	unsigned written = 0;
-	unsigned sent = 0;
-};
-
-class FtpDataFileList : public FtpDataStream
-{
-public:
-	explicit FtpDataFileList(FtpServerConnection* connection) : FtpDataStream(connection)
-	{
-	}
-
-	void transferData(TcpConnectionEvent sourceEvent) override
-	{
-		if(completed) {
-			return;
-		}
-		Vector<String> list = fileList();
-		debug_d("send file list: %d", list.count());
-		for(unsigned i = 0; i < list.count(); i++) {
-			writeString("01-01-15  01:00AM               " + String(fileGetSize(list[i])) + " " + list[i] + "\r\n");
-		}
-		completed = true;
-		finishTransfer();
-	}
-};
-
-class FtpDataRetrieve : public FtpDataStream
-{
-public:
-	FtpDataRetrieve(FtpServerConnection* connection, const String& fileName) : FtpDataStream(connection)
-	{
-		file = fileOpen(fileName, eFO_ReadOnly);
-	}
-
-	~FtpDataRetrieve()
-	{
-		fileClose(file);
-	}
-
-	void transferData(TcpConnectionEvent sourceEvent) override
-	{
-		if(completed) {
-			return;
-		}
-		char buf[1024];
-		int len = fileRead(file, buf, 1024);
-		write(buf, len, TCP_WRITE_FLAG_COPY);
-		if(fileIsEOF(file)) {
-			completed = true;
-			finishTransfer();
-		}
-	}
-
-private:
-	file_t file;
-};
-
-class FtpDataStore : public FtpDataStream
-{
-public:
-	FtpDataStore(FtpServerConnection* connection, const String& fileName) : FtpDataStream(connection)
-	{
-		file = fileOpen(fileName, eFO_WriteOnly | eFO_CreateNewAlways);
-	}
-
-	~FtpDataStore()
-	{
-		fileClose(file);
-	}
-
-	virtual err_t onReceive(pbuf* buf)
-	{
-		if(completed) {
-			return TcpConnection::onReceive(buf);
-		}
-
-		if(buf == nullptr) {
-			completed = true;
-			response(226, "Transfer completed");
-			return TcpConnection::onReceive(buf);
-		}
-
-		pbuf* cur = buf;
-		while(cur != nullptr && cur->len > 0) {
-			fileWrite(file, (uint8_t*)cur->payload, cur->len);
-			cur = cur->next;
-		}
-
-		return TcpConnection::onReceive(buf);
-	}
-
-private:
-	file_t file;
-};
-
-///////////////////////////////////////////////////////////////////////////////////////////////////////////
+//#include "../FtpServer.h"
+//#include "NetUtils.h"
+//#include "TcpConnection.h"
+#include "FtpDataStore.h"
+#include "FtpDataRetrieve.h"
+#include "FtpDataFileList.h"
 
 err_t FtpServerConnection::onReceive(pbuf* buf)
 {
